Change type of `Network.Pcap.loop*` to return unit.
And fix documentation for the `loop*` functions. The documentation
wrongly said they return the number of packets read. They actually
return 0 on success, -1 on error, and -2 if interrupted in a special
way by `pcap_breakloop`. We don't expose the `pcap_breakloop` API, and
we raise an error on -1, so these functions will always return 0 if
used via our APIs. I left the `Int` return type on the lower level
`Network.Pcap.Base.loop` function, in case someone happens to
implement the `pcap_breakloop` API in their own code.