Allow Ctrl-C to interrupt a pcap dump loop without corruption.
I had noticed that killing my program with Ctrl-C often resulted in
corrupted dump files when I ran `loop` with a callback to `dump`. So,
I changed the `ForeignPtr` creation in `openDump` to use a version of
`newForeignPtr` that guarantees that the supplied finalizer will be
run.

Related comment I added to `openDump`:

    -- Using 'Foreign.Concurrent.newForeignPtr' isn't sufficient
    -- here, since its (Haskell) finalizer is not guaranteed to
    -- run. Rather, we use 'Foreign.ForeignPtr.newForeignPtr', since
    -- its (C) finalizer is guaranteed to run. Previously, when
    -- using 'Foreign.Concurrent.newForeignPtr' and killing the
    -- program with Ctrl-C, we sometimes got corrupted dump files.
    --
    -- Perhaps we should also change other finalizers -- 'openDead',
    -- 'openLive', 'openOffline', 'compileFilter' -- to use
    -- 'Foreign.ForeignPtr.newForeignPtr'? But handling the "whole
    -- program terminated unexpectedly" case seems unimportant for
    -- those, since there doesn't seem to be a corruption concern.

Note that attempting to break a pcap loop by killing it's
thread (e.g. with `Async.cancel` or `throwTo`) still doesn't work! I
tried using the `InterruptibleFFI` extention [1] to make the FFI call
to `pcap_loop` interruptible, but that didn't have any effect. Since
we want to be able to cancel threads running pcap loops, we probably
want to expose `pcap_break_loop`.

[1] InterruptibleFFI extension:
    https://downloads.haskell.org/~ghc/8.2.1/docs/html/users_guide/ffi-chap.html#ffi-interruptible